Improved borg item selection behavior
Behavior changed to:
Pressing X to cycle slots goes to next slot with an active module in it.
If at slot 3, it will deselect all slots.
if initial slot has nothing in it it will attempt slot 2, or 3
if slot 2 is empty, but 1 and 3 have something it will rotate 1 to 3 to no slots
this behavior now always attempts to select at least something, and if at the last slot deselects all slots to allow easy interaction with some map objects as borgs